chore:#02 canvasをとって渡す処理をなまぶ
