+++ src/blur_debug/scripts/bootstrap.js
@@ -3,9 +3,12 @@ import world from "./glsl/world";
 import scroller from "./component/scroller";
 import mouse from "./component/mouse";
 import viewport from "./helper/viewport";
+import { iNode } from "../../iNode";
 
 export function init() {
-  viewport.init();
+  const canvas = iNode.qs("#canvas");
+
+  viewport.init(canvas);
 
   world.init();
